Consolidate tmpl data .go files.
Merge data_rootdir.go into data.go. It doesn't need to be a standalone
file. This is consistent with other uses of vfsgen in related packages.
